move tool test code located in pkg to outside
